--- js/length-line.js
@@ -1,10 +1,0 @@
-// Функция для проверки длины строки.
-
-function checkStringLenght(string,maxLength){
-  if(string.length <= maxLength){
-    return true;
-  }
-  return false;
-}
-
-export {checkStringLenght};
